requirements.txt
absl-py==1.0.0
astor==0.8.1
Box2D==2.3.10
cached-property==1.5.2
certifi==2020.12.5
cffi==1.15.0
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
gast==0.2.2
glfw==2.4.0
google-pasta==0.2.0
grpcio==1.41.1
gym==0.21.0
h5py==3.1.0
imageio==2.10.4
importlib-metadata==4.8.2
importlib-resources==5.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.4.58
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.1
pycparser==2.21
pygame==2.1.0
pyglet==1.5.21
PyOpenGL==3.1.5
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==4.0.0
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
zipp==3.6.0
